+++ pytzer/__init__.py
@@ -1,4 +1,4 @@
-# Pytzer: Pitzer model for chemical activities in aqueous solutions
+# Pytzer: Pitzer model for chemical activities in aqueous solutions.
 # Copyright (C) 2019  Matthew Paul Humphreys  (GNU GPLv3)
 #
 # This program is free software: you can redistribute it and/or modify
@@ -13,11 +13,9 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
 """Pitzer model for chemical activities in aqueous solutions."""
-
-from . import cflibs, coefficients, constants, debyehueckel, io, jfuncs, \
-    matrix, meta, model, properties, tables, teos10
+from . import (cflibs, coefficients, constants, debyehueckel, io, jfuncs,
+    matrix, meta, model, properties, tables, teos10)
 
 __all__ = [
     'cflibs',
@@ -75,7 +73,7 @@ def blackbox(filename, cflib=cflibs.Seawater, savefile=True):
     acfs[:,L] = model.acfs(*nargsL)
     if np_any(~L):
         acfs[:, ~L] = model.acfs(*nargsLx, Izero=True)
-    # Save results unless requested not to
+    # Save results, unless requested not to
     if savefile:
         filestem = filename.replace('.csv','')
         io.saveall(filestem + '_py.csv',