+++ pytzer/cflibs.py
@@ -1,8 +1,6 @@
-# Pytzer: Pitzer model for chemical activities in aqueous solutions
+# Pytzer: Pitzer model for chemical activities in aqueous solutions.
 # Copyright (C) 2019  Matthew Paul Humphreys  (GNU GPLv3)
-
 """Assemble dicts of Pitzer model coefficient functions."""
-
 from . import debyehueckel, jfuncs, properties
 from . import coefficients as cf
 from .meta import version