+++ pytzer/meta.py
@@ -1,6 +1,5 @@
-# Pytzer: Pitzer model for chemical activities in aqueous solutions
+# Pytzer: Pitzer model for chemical activities in aqueous solutions.
 # Copyright (C) 2019  Matthew Paul Humphreys  (GNU GPLv3)
-
 """Define module metadata."""
 
-version = '0.3.0.1'
+version = '0.3.1'